+++ src/main/java/com/gymcrm/gymcrm/repository/TrainingRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface TrainingRepository extends JpaRepository<Training, Long> {
+public interface TrainingRepository extends JpaRepository<Training, Integer> {
     List<Training> findTrainingsByTrainerAndTrainee(Trainer trainer, Trainee trainee);
 }
